requirements.txt
beautifulsoup4==4.14.3
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
curl_cffi==0.13.0
fastapi==0.115.0
frozendict==2.4.7
idna==3.11
Jinja2==3.1.5
llvmlite==0.44.0
multitasking==0.0.12
numba==0.61.2
numpy==2.2.6
pandas==2.3.3
pandas-ta==0.4.71b0
peewee==3.18.3
platformdirs==4.5.1
protobuf==6.33.2
pycparser==2.23
python-multipart==0.0.12
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
six==1.17.0
soupsieve==2.8.1
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.2
uvicorn==0.30.6
websockets==15.0.1
yfinance==1.0
plotly==5.24.1
